Rename package gg.jte.nullmarked to gg.jte.jspecify
Addresses reviewer feedback: aligns Java package name with the module
name (jte-jspecify-nullmarked).
